+++ FirebaseChat_2017_09_07/app/build.gradle
@@ -22,7 +22,7 @@ android {
 dependencies {
     dependencies {
         compile fileTree(dir: 'libs', include: ['*.jar'])
-        testCompile 'junit:junit:4.12'
+
         compile 'com.android.support:appcompat-v7:23.2.1'
 
         compile 'com.google.firebase:firebase-core:9.0.0'
@@ -33,7 +33,7 @@ dependencies {
         compile 'com.android.support.constraint:constraint-layout:1.0.2'
         compile 'com.google.firebase:firebase-messaging:9.0.0'
         compile 'com.firebase:firebase-jobdispatcher:0.7.0'
-
+        testCompile 'junit:junit:4.12'
     }
 }
 apply plugin: 'com.google.gms.google-services'
